backend/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic_settings==2.3.1
